+++ src/main/java/net/aoba/gui/tabs/components/ButtonComponent.java
@@ -27,9 +27,9 @@ public class ButtonComponent extends Component {
 	}
 
 	public void update(int offset, double mouseX, double mouseY, boolean mouseClicked) {
-		int parentX = parent.getX();
-		int parentY = parent.getY();
-		int parentWidth = parent.getWidth();
+		float parentX = parent.getX();
+		float parentY = parent.getY();
+		float parentWidth = parent.getWidth();
 
 		if (HudManager.currentGrabbed == null) {
 			if (mouseClicked) {
@@ -66,9 +66,9 @@ public class ButtonComponent extends Component {
 
 	@Override
 	public void draw(int offset, DrawContext drawContext, float partialTicks, Color color) {
-		int parentX = parent.getX();
-		int parentY = parent.getY();
-		int parentWidth = parent.getWidth();
+		float parentX = parent.getX();
+		float parentY = parent.getY();
+		float parentWidth = parent.getWidth();
 		MatrixStack matrixStack = drawContext.getMatrices();
 		renderUtils.drawOutlinedBox(matrixStack, parentX + 2, parentY + offset, parentWidth - 4, this.getHeight() - 2,
 				backgroundColor, 0.2f);
